+++ NEWS.md
@@ -4,6 +4,7 @@ vXXX - unreleased
 * fixed: parsing IPv6 addresses, or addresses in the form `[127.0.0.1]:8080` in `pinba_server` configuration option
 * improved: added support for `pinba.auto_flush` configuration option
 * improved: report automatically to Pinba the script's http status code by default
+* improved: support custom `hit_count` on timer creation
 * improved: replicate extension behaviour: default `req_count` is 1 in data from `get_info()`, but 0 as sent to the server
 * improved: added one more sample file: doc/measure_body_size.php